OK lesson 78

+++ app/Http/Controllers/Admin/RolesController.php
@@ -17,6 +17,7 @@ class RolesController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', new Role);
         return view('admin.roles.index', [
             'roles'=>Role::all()
         ]);
@@ -29,10 +30,10 @@ class RolesController extends Controller
      */
     public function create()
     {
-
+        $this->authorize('create', $role = new Role);
         return view('admin.roles.create', [
             'permissions'=>Permission::pluck('name', 'id'),
-            'role'=>new Role,
+            'role'=>$role,
         ]);
     }
 
@@ -52,24 +53,13 @@ class RolesController extends Controller
         //     'name.unique'=>'Este campo identificador já está em uso.',
         //     'display_name.required'=>'O campo nome é obrigatório.'
         // ]);
+        $this->authorize('create', new Role);
         $role = Role::create($request->validated());
         if ($request->has('permissions')) {
             $role->givePermissionTo($request->permissions);
         }
         return redirect()->route('admin.roles.index')->withFlash('O papel foi criado corretamente.');
     }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -78,6 +68,7 @@ class RolesController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('update', $role);
         return view('admin.roles.edit', [
             'role'=>$role,
             'permissions'=>Permission::pluck('name', 'id'),
@@ -97,6 +88,7 @@ class RolesController extends Controller
         //     [
         //         'display_name.required'=>'O campo nome é obrigatório.'//se falhar acima, exibir msg
         //     ]);
+        $this->authorize('update', $role);
         $role->update($request->validated());
         $role->permissions()->detach();
         if ($request->has('permissions')) {
@@ -113,10 +105,7 @@ class RolesController extends Controller
      */
     public function destroy(Role $role)
     {
-        if ($role->id === 1) {
-            throw new \Illuminate\Auth\Access\AuthorizationException('Não é possível eliminar este papel!');
-        }
-        //$this->authorize('delete', $role);
+        $this->authorize('delete', $role);
         $role->delete();
         return redirect()->route('admin.roles.index')->withFlash('O papel foi eliminado');
     }
